projectDetailをproductDetailに修正
